extract FetcherCallable and ProjectContributorFetcherFunction and added a junit test
